Rework how return messages are processed
Instead of writing return messages out to the user in our process methods,
just return the method to main so it can be written there. This allows us to
easier handle the cases where our socket goes bad on us and we can no longer
write to it.

Also fix a small off-by-one error in our input parser, along with a stupid
issue where we passed a connection on the stack rather than a pointer to a
connection, causing us to lose all state if we ever had characters left over
in the buffers.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>