Add logic to control program to wait between sending commands
Ensure the receiver has time to handle the commands we are sending it by
implementing a wait time between sends if we have many commands in the
queue. It seems to behave better this way, especially with back to back
status and zone 2 status commands where a lot of stuff is being rammed down
the serial line.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>