Add a command sending delay
Do some funky stuff in our select loop buildup so we only send a command if
it has been a minimum amount of time since the previous command was sent.
This should ensure we don't overload the receiver and have dropped commands
like we were seeing after sending back-to-back 'status' and 'status zone2'
commands.

The select() loop now gets a timeout value if we determine it hasn't been
long enough since our last write to the serial device port. The timeout is
set to fire once enough time has elapsed, which allows our write file
descriptor to be re-added the next time around and the next command written
to the receiver. In cases where there are no commands in the queue, the
select() will act as before and block until something is available to read.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>