Grey out controls that cannot be used when appropriate
If either the main zone power or zone 2 power is off, we can grey quite a
few of our shown controls as they will not have an effect. Do this, and at
the same time, make the update function a bit more efficient by only
updating what changed on us.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>