Prevent queueing of duplicate commands
When queueing commands to send to the receiver, don't add commands that are
already present in the queue. This would prevent the case where 3 frontends
all ask for status and we really only need one set of feedback from the
receiver.

There is one drawback here- a sequence such as "mute on", "mute off", and
"mute on" would drop the last "mute on" command and leave it in the unmuted
state.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>