Add power state logic to backend controller
This should resolve a lot of the issues with overloading the receiver with
commands when it is off. By keeping track of receiver state, we prevent
commands from being sent that don't have to deal with power when the power
is off. Some bitmask logic allows us to keep track of each zones' power
state as well, so we let commands through if zone 2 is on even if the
primary power is off.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>