Use memmem instead of strstr in message parsing
I noticed some reads coming in like the following using strace:
read(5, "\0\0!1LMD81\32"..., 63)        = 10

We then proceeded to strip off the two leading bytes (the size of
START_RECV), but this left the "!1" string on the status message so we did
not successfully interpret it. By using memmem() to search for our start
character sequence, we should be able to ignore any null or other garbage
and go straight to where we want to go.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>