Remove arbitrary MAX_* restrictions
Move listeners and connections to variable length linked lists instead of
fixed compile-time sized arrays. This gives everything a bit more
flexibility at the cost of a bit more code, but overall it should be a good
step to take. Buffers and structures are allocated on an as-needed basis,
and once they are allocated they will be reused until program exit. There is
still an upper limit on connections at 200 to prevent any kind of DoS
attack, but it could really be raised as high as one wants, and the select()
call is going to be the first thing to topple.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>