Add ability to listen on UNIX socket
Most of the credit for this goes elsewhere, but I worked it into the
now-existent option parsing framework and listening code. Note that you can
listen on both a UNIX socket and an INET one; there is also only minor
changes necessary to support listening on arbitrary numbers of both.

Original-work-by: Jasmin Buchert <jbuchert@gmail.com>
Signed-off-by: Dan McGee <dpmcgee@gmail.com>