Receiver global state refactor
This is in anticipation of doing the eISCP network-based control, but I
haven't been too happy with the amount of global state for a while. Get it a
bit more wrapped up in this receiver class, which whether we need it or not,
now allows us to send commands to as many receivers as we want.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>