Push write_to_connections() calls deeper into the stack
We used to only call write_to_connections in the main loop. Now that
this function has been exposed anyway for use by the zone sleep code, it
makes more sense for this to be called directly in the response parsing
code as necessary. This allows us to skip silly allocation for the
response strings, as well as opens up the possibility of having a single
message from the receiver turn into multiple response messages. We take
advantage of this by adding 'OK:dbvolume:-44' type responses in addition
to the 'OK:volume:38' messages already there.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>