Treat receiver status messages the same way we do commands
Allocate a non-const set of structs in place, and simply hash the names
at initialization time in place. This prevents the need for any memory
allocation and thus teardown.

This separates the power statuses from the rest so we can attach some
metadata; this will come into play in a future commit that reworks the
way we do update_power_status().

Signed-off-by: Dan McGee <dpmcgee@gmail.com>