Fix some faulty timeval related math and comparisons
Add a timeval_positive helper function that doesn't get confused when
tv_usec is set to 0 (regardless of the value of tv_sec). Use it wherever
possible, and also attempt to reason through our other time value comparison
points to ensure all values are being handled correctly, especially those
that result in a zero in a field or a zero difference.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>